the right wing gets involved ;)
meaning we are starting to use the right stick (exclusively for sound) and the left stick only deals with the render process, this is only to ensure we can separate the two sticks during the platform-game layer transition. This is working as expected. (Still not sure about the whole OldController->IsAnalog thing).